refactor(foldkit): move the hydration handoff out of the boot closure
`startWith` decided inline whether a boot adopts the server-rendered
root and where its Flags come from: the fallback to the app's own Flags
Effect, the refusal that contains the page when a hydrating boot finds
no stamped root, the build id comparison, the choice to skip adoption
for an HMR-restored Model, and the lookup and decode of the page's Flags
payload with its own refusals.

`resolveHydrationHandoff` in `hydrationHandoff.ts` now owns all of that,
next to `buildSkew` and `containRefusedPage`, which it calls. It takes
the boot mode, the hydration config, the boot and configured Flags
Effects, `isFlagsRequired`, the Flags codec, the HMR Model, the
container, the build id, and `provideResources`, and returns the root
the first render adopts, when there is one, and the Effect that
produces the Flags `init` runs with. Each refusal still ends startup
the same way, from inside the unit, so the boot stops where it did.

The boot calls it right after the resource provider, where the block
was, and nothing between that point and `init` changed. `BootMode`
moved to `hydrationHandoff.ts` with the unit, since that is what reads
it, and `start.ts` imports it from there. The moved code is unchanged
apart from formatting and the parameter types it gained.

No user-facing effect, so the changeset is empty.
